@@ -18,18 +18,18 @@
   "dependencies": {
     "@akveo/ng2-completer": "^9.0.1",
     "@angular-devkit/core": "^16.2.8",
-    "@angular/animations": "^13.4.0",
+    "@angular/animations": "^14.3.0",
     "@angular/cdk": "12.1.0",
-    "@angular/cli": "^13.3.11",
-    "@angular/common": "^13.4.0",
-    "@angular/compiler": "^13.4.0",
-    "@angular/compiler-cli": "^13.4.0",
-    "@angular/core": "^13.4.0",
-    "@angular/forms": "^13.4.0",
+    "@angular/cli": "^14.2.13",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
     "@angular/google-maps": "^12.1.0",
-    "@angular/platform-browser": "^13.4.0",
-    "@angular/platform-browser-dynamic": "^13.4.0",
-    "@angular/router": "^13.4.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@asymmetrik/ngx-leaflet": "3.0.1",
     "@nebular/auth": "8.0.0",
     "@nebular/eva-icons": "8.0.0",
@@ -73,8 +73,8 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.11",
-    "@angular/language-service": "13.4.0",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular/language-service": "14.3.0",
     "@compodoc/compodoc": "1.0.1",
     "@fortawesome/fontawesome-free": "^5.2.0",
     "@types/d3-color": "1.0.5",
