@@ -1,5 +1,5 @@
 import { ChangeDetectorRef, Component, Inject, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { NbAuthService, NbLoginComponent, NB_AUTH_OPTIONS } from '@nebular/auth';
 import { ApiService } from '../../../../service/auth/api.service';
@@ -11,7 +11,7 @@ import { ApiService } from '../../../../service/auth/api.service';
 })
 export class LoginComponent extends NbLoginComponent {
 
-  constructor(private fb: FormBuilder, private apiService: ApiService, service: NbAuthService, @Inject(NB_AUTH_OPTIONS) options: {}, cd: ChangeDetectorRef, router: Router) {
+  constructor(private fb: UntypedFormBuilder, private apiService: ApiService, service: NbAuthService, @Inject(NB_AUTH_OPTIONS) options: {}, cd: ChangeDetectorRef, router: Router) {
     super(service, options, cd, router);
   }
 
